pyproject.toml
[tool.poetry]
name = "brower-tool"
version = "0.1.0"
description = "browser_use with remote browser"
authors = ["roloritun <rahmanoloritun@outlook.com>"]
readme = "README.md"
packages = [
    {include = "src"},
    {include = "docker/browser"}
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
browser-use = ">=0.1.48,<0.2.0"
daytona-sdk = ">=0.17.0,<0.18.0"
pytest = ">=8.3.5,<9.0.0"
langchain = ">=0.3.21"
langchain-openai = ">=0.3.11"
python-dotenv = ">=1.1.0"
requests = ">=2.31.0"
langchain-core = ">=0.3.49"
pydantic = ">=2.0.0"
aiohttp = "^3.11.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
